Implemento IDispatcheable

+++ src/main/kotlin/com/thegrid/behavior/platform/IDispatcheable.kt
@@ -5,6 +5,6 @@ interface IDispatcheable {
     /**
      * @return lo que duró el evento
      */
-    fun executeEvent() : Long
+    fun executeEvent() : Double
     
 }
